requirements.txt
numpy>=1.26.0
python-dotenv>=1.0.1
transformers>=4.46.0
torch>=2.2.0
accelerate>=0.34.0
